memset

+++ srcs/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: olivier <olivier@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2013/11/19 14:40:55 by ocosquer          #+#    #+#             */
-/*   Updated: 2015/10/07 14:26:27 by olivier          ###   ########.fr       */
+/*   Updated: 2015/10/08 13:58:01 by olivier          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 void	*ft_memset(void *byte, int c, size_t len)
 {
-	char			new_char;
-	char			*new_array;
-	unsigned int	length;
+	char		new_char;
+	char		*new_array;
+	size_t		length;
 
 	length = 0;
 	new_char = (char)c;
